+++ lib/rubygems/command.rb
@@ -430,9 +430,9 @@ class Gem::Command
 
   def handles?(args)
     parser.parse!(args.dup)
-    return true
+    true
   rescue StandardError
-    return false
+    false
   end
 
   ##
@@ -459,7 +459,7 @@ class Gem::Command
     until extra.empty? do
       ex = []
       ex << extra.shift
-      ex << extra.shift if extra.first.to_s =~ /^[^-]/ # rubocop:disable Performance/StartWith
+      ex << extra.shift if /^[^-]/.match?(extra.first.to_s) # rubocop:disable Performance/StartWith
       result << ex if handles?(ex)
     end