+++ lib/rubygems/config_file.rb
@@ -202,7 +202,7 @@ class Gem::ConfigFile
       @hash = @hash.merge environment_config
     end
 
-    # HACK these override command-line args, which is bad
+    # HACK: these override command-line args, which is bad
     @backtrace                   = @hash[:backtrace]                   if @hash.key? :backtrace
     @bulk_threshold              = @hash[:bulk_threshold]              if @hash.key? :bulk_threshold
     @home                        = @hash[:gemhome]                     if @hash.key? :gemhome
@@ -479,7 +479,7 @@ if you believe they were disclosed to a third party.
 
     @hash.each do |key, value|
       key = key.to_s
-      next if key =~ re
+      next if key&.match?(re)
       yaml_hash[key.to_s] = value
     end
 
@@ -581,7 +581,7 @@ if you believe they were disclosed to a third party.
         need_config_file_name = false
       elsif arg =~ /^--config-file=(.*)/
         @config_file_name = $1
-      elsif arg =~ /^--config-file$/
+      elsif /^--config-file$/.match?(arg)
         need_config_file_name = true
       end
     end