+++ lib/rubygems/doctor.rb
@@ -111,7 +111,7 @@ class Gem::Doctor
 
       basename = File.basename(child, extension)
       next if installed_specs.include? basename
-      next if /^rubygems-\d/ =~ basename
+      next if /^rubygems-\d/.match?(basename)
       next if "specifications" == sub_directory && "default" == basename
       next if "plugins" == sub_directory && Gem.plugin_suffix_regexp =~ (basename)