+++ lib/rubygems/package.rb
@@ -187,7 +187,7 @@ class Gem::Package
       end
     end
 
-    return spec, metadata
+    [spec, metadata]
   end
 
   ##